requirements.txt
discord.py>=2.6.2
aiohttp>=3.9.0
aiofiles
discord.py